fix(capabilities): target the listener address for the rpc server teardown self-connect
Extract #2631's teardown_connect_addr helper into shared::net and reuse
it in the rpc server's unwire, replacing the hardcoded 127.0.0.1
self-connect (and its swallowed connect error) with the listener's
actual bound address plus a warn! on failure.

Closes #2645
