+++ crates/aether-capabilities/src/shared/mod.rs
@@ -1,3 +1,5 @@
 //! Shared infrastructure consumed by multiple capabilities but not itself a capability.
 #[cfg(not(target_family = "wasm"))]
 pub mod contentgen;
+#[cfg(not(target_family = "wasm"))]
+pub mod net;
